Add caching system
Prompt: "Add a cache.py file, which implements a key value store on the disk with read and write. Values should be stored in a .cache/ directory, using md5 hash of the keys. Reading a non existent key should throw an exception.

Add a function which can memoize function calls by storing their arguments to the KV store."