Add command.py
Prompt: "Add a new file, command.py. It is going to parse and pretty print commands.

It should have a function to turn a command into a printed string. It should have a function to parse a string, which returns an *array of commands*, as there may be multiple. The parse function does not need to support commands with responses.

Commands in Python are stored as dicts. The "command" key stores the id of the key. The optional "body" key stores the body of the command. The optional "response" key contains the other system's response to the command. All other keys are a key/value pair, which are all string/strings.

In text, the format of a command is as follows:

@@COMMAND_ID@@ key_a=value_a key_b=value_b
OPTIONAL
BODY
OVER
MULTIPLE
LINES

If the command has a response, it should be printed with the command quoted forum style, with the response below:

> @@COMMAND_ID@@ key_a=value_a key_b=value_b
> OPTIONAL
> BODY
> OVER
> MULTIPLE
> LINES
RESPONSE LINE A
RESPONSE LINE B"