Simplify Command
Prompt: "Modify command.py to remove the concept of "response". Document the protocol at the top of the file in a long comment.

The spec of the file is now: 

It should have a function to turn a command into a printed string. It should have a function to parse a string, which returns an array of commands, as there may be multiple. The parse function does not need to support commands with responses.

Commands in Python are stored as dicts. The "command" key stores the id of the key. The optional "body" key stores the body of the command. All other keys are a key/value pair, which are all string/strings.

In text, the format of a command is as follows:

@@COMMAND_ID@@ key_a=value_a key_b=value_b
OPTIONAL
BODY
OVER
MULTIPLE
LINES"