Add search tool
Prompt: "Add a search tool to tools/search.py. It will take a string, and return the filenames that contain that string.

The parameters should be a files dict from issue.py, and a search string. It should return a list of all of the keys where the value contains the search string."