Expand Command schema
Prompt: "In commands/command.py, the schema functions should return the full schema for the command, as documented at the top of the file. Right now they just document the parameters.

The names of the commands are just the names of the classes.

The description for Think is: "Write out a plan for the steps you are about to take. Please think before taking any action. Include at least a paragraph of detail"

The description for Verdict is: "Make a decision on whether the proposed change is acceptable or not""