Optimise Tokens
Prompt: "In command_files.py, modify the files that are written out during execute. Do this before adding the line numbers.

For each line, detect how many spaces the line starts with. If it's a multiple of 4, then replace each group of 4 spaces with a tab. 

For instance, if it starts with 12 spaces, which represents an indentation of 3 tabs, replace the line with 3 tabs and then the rest of the line.

Write this logic as a separate function.
"