Move MAX_WORKERS into Settings object
Prompt: "Title: Move MAX_WORKERS into Settings object
Description: In settings.py, move MAX_WORKERS into the Settings object, as a max_workers member variable. 

Access to this setting must be via get_settings(). Don't add a getter for the member variable, and leave all of the other settings as is for now."