Add a move_file_using_rope function
Prompt: "Title: Add a move_file_using_rope function
Description: Take a project directory, as well as a relative from and a relative to path. 

Add it to move_file.py. 

You must the rope library's Rename tool, under rope.refactor.rename, using the file as a resource. In Rope, Rename on a file renames the file on disk. Rope's Move command is not intended to move files.

After you have a changeset, project.do(changes) will run it."